chore/update: update app
