+++ graphql-api/src/Lease/Lease.Fields.fs
@@ -98,6 +98,7 @@ let ListLeaseEventsInputObject =
         description = "Input for listing lease events",
         fields = [
             leaseIdInputField
+            asOfDateInputField
             pageSizeInputField
             pageTokenInputField
         ]
