+++ web-app/package.json
@@ -1,19 +1,25 @@
 {
-  "name": "web",
+  "name": "tutorial",
   "version": "0.1.0",
   "private": true,
+  "devDependencies": {
+    "@types/jest": "^24.0.15",
+    "@types/node": "^12.0.10",
+    "@types/react": "^16.8.22",
+    "@types/react-dom": "^16.8.4",
+    "@types/uuid": "^3.4.4"
+  },
   "dependencies": {
     "@apollo/react-hooks": "^3.1.1",
     "@date-io/date-fns": "^1.3.7",
     "@material-ui/core": "^4.4.3",
+    "@material-ui/icons": "^4.4.3",
     "@material-ui/styles": "^4.4.3",
-    "@types/jest": "24.0.15",
-    "@types/node": "12.0.10",
-    "@types/react": "16.8.22",
-    "@types/react-dom": "16.8.4",
-    "@types/uuid": "^3.4.4",
-    "apollo-boost": "^0.4.4",
+    "apollo-cache-inmemory": "^1.6.3",
+    "apollo-client": "^2.6.4",
+    "apollo-link-http": "^1.5.16",
     "graphql": "^14.4.0",
+    "graphql-tag": "^2.10.1",
     "material-table": "^1.40.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
