+++ web-app/src/components/Root.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import { Root } from './Root';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+  ReactDOM.render(<Root />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
